Moved dashboard page to entry page
Dashboard logo + dashboard item (image/text) now loads with http://localhost:3000/

(instead of http://localhost:3000/dashboard)
